src/requirements.txt
langchain
fastapi
uvicorn
pydantic
langchain-google-genai
python-dotenv
youtube-transcript-api
langchain_community
pinecone
langchain_pinecone
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart